+++ app/renderer/src/components/result-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './result-list';
