+++ app/renderer/src/containers/result-item-container/index.js
@@ -1,0 +1,1 @@
+export { default } from './result-item-container';
