+++ app/renderer/src/containers/result-list-container/index.js
@@ -1,0 +1,1 @@
+export { default } from './result-list-container';
